Turn project list into component
