+++ pages/index.jsx
@@ -1,20 +1,9 @@
 import Head from 'next/head';
-import Link from 'next/link';
-import Layout, { siteTitle } from '../components/layout';
-import Date from '../components/date';
+import { siteTitle } from '../components/layout';
+import ProjectList from '../components/projectList';
 import utilStyles from '../styles/utils.module.scss';
-import { getSortedProjectsData } from '../lib/projects';
 
-export async function getStaticProps() {
-  const allProjectsData = getSortedProjectsData();
-  return {
-    props: {
-      allProjectsData,
-    },
-  };
-}
-
-export default function Home({ allProjectsData }) {
+export default function Home() {
   return (
     <Layout home>
       <Head>
@@ -23,27 +12,7 @@ export default function Home({ allProjectsData }) {
       <section className={utilStyles.headingMd}>
         <p className='text-center'>aspiring software engineer and marketing coordinator</p>
       </section>
-      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
-        <div>
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="max-w-2xl mx-auto py-16 sm:py-24 lg:py-32 lg:max-w-none">
-
-              <div className="mt-6 space-y-12 lg:space-y-0 lg:grid lg:grid-cols-3 lg:gap-x-6">
-                {allProjectsData.map(({ id, title, date }) => (
-                  <div key={id} className="group relative">
-                    <h3 className="mt-6 text-xl text-gray-500">
-                      <Link href={`/projects/${id}`}>
-                        {title}
-                      </Link>&nbsp;
-                      <Date dateString={date} />
-                    </h3>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      <ProjectList />
     </Layout>
   );
 }
